Restructure nodes.
Moved base nodes to the top level, the derivative ones are in their own
categories.
Also renamed some stuff.
