+++ src/nodes/converter/blackbody.c
@@ -1,0 +1,9 @@
+//
+//  blackbody.c
+//  C-Ray
+//
+//  Created by Valtteri Koskivuori on 16/12/2020.
+//  Copyright © 2020 Valtteri Koskivuori. All rights reserved.
+//
+
+#include "blackbody.h"
