+++ src/nodes/converter/grayscale.h
