+++ src/nodes/shaders/diffuse.c
@@ -11,11 +11,11 @@
 #include "../../renderer/samplers/sampler.h"
 #include "../../datatypes/vector.h"
 #include "../../datatypes/material.h"
-#include "../textures/texturenode.h"
+#include "../colornode.h"
 #include "../../utils/hashtable.h"
 #include "../../datatypes/scene.h"
 #include "../../utils/logging.h"
-#include "bsdf.h"
+#include "../bsdf.h"
 
 #include "diffuse.h"
