+++ src/nodes/textures/image.c
@@ -16,7 +16,7 @@
 #include "../../datatypes/hitrecord.h"
 #include "../../utils/hashtable.h"
 #include "../../datatypes/scene.h"
-#include "texturenode.h"
+#include "../colornode.h"
 
 #include "image.h"
