+++ src/nodes/vectornode.c
@@ -1,0 +1,15 @@
+//
+//  vectornode.c
+//  C-Ray
+//
+//  Created by Valtteri Koskivuori on 16/12/2020.
+//  Copyright © 2020 Valtteri Koskivuori. All rights reserved.
+//
+
+#include "../includes.h"
+#include "../datatypes/color.h"
+#include "../datatypes/vector.h"
+#include "../datatypes/hitrecord.h"
+#include "bsdf.h"
+
+#include "vectornode.h"
